Extracted some functions into helpers/utils, addressed PR comments  Please enter the commit message for your changes. Lines starting
