+++ setup.cfg
@@ -53,3 +53,4 @@ addopts = --verbose -rsxX -q
 testpaths = tests
 markers =
 	gfxtest: marks a test as needing to render
+norecursedirs=tests/helpers
