+++ tests/conftest.py
@@ -1,4 +1,9 @@
+import os
 from os import path as osp
+import sys
+sys.path.append(os.path.join(os.path.dirname(__file__), 'helpers'))
+
+
 
 import pytest
